requirements.txt
Django==4.2
psycopg==3.1.17
djangorestframework==3.14.0
phonenumbers==8.13.30
django-phonenumber-field==7.3.0
